Rename RadicleApplicationService to RadicleProjectService
